+++ airflow/serialization/serialized_objects.py
@@ -58,7 +58,7 @@ BUILTIN_OPERATOR_EXTRA_LINKS: List[str] = [
     "airflow.providers.google.cloud.operators.bigquery.BigQueryConsoleIndexableLink",
     "airflow.providers.google.cloud.operators.mlengine.AIPlatformConsoleLink",
     "airflow.providers.qubole.operators.qubole.QDSLink",
-    "airflow.operators.dagrun_operator.TriggerDagRunLink",
+    "airflow.operators.trigger_dagrun.TriggerDagRunLink",
     "airflow.sensors.external_task_sensor.ExternalTaskSensorLink",
 ]
